Make some changes to our Vagrant portability testing to be more amenable
to automation via a Jenkins job.

Changes:
- Added a provisioning step which runs a script to propagate
`http[s]_proxy` env vars into the VM permanently, so they function on
systems behind a proxy.
- Make VM management scripts operate on a single VM by default,
controlled by `VM_NAME` env var.
- Add separate versions that always operate on all VMs, like
`up-all.sh`.
- Also updated to testing only the currently supported FreeBSD releases
while here.
- Use the same amount of resources for every VM.
- Fix relative working directory issues in some provision scripts.

To be merged alongside a CI change adding such a Jenkins job.

Resolves https://github.com/Cray/chapel-private/issues/2000.

[reviewed by @jabraham17, thanks!]

Testing:
- [x] manual run of test job
- [x] put `git-clone-chapel.sh` back to cloning `main` before merge

+++ util/devel/test/portability/vagrant/destroy.sh
@@ -4,10 +4,23 @@
 # Removes space occupied by each VM
 # Don't do this if you have work you want to save in the VMs!
 
+if [ -z "$VM_NAME" ]
+then
+  echo "Please set the VM_NAME environment variable to the name of the VM you want to run this on."
+  exit 1
+fi
+
 for name in current/*
 do
   if [ -f $name/Vagrantfile ]
   then
+    if [ "$VM_NAME" != "all" ]
+    then
+      if [ "$name" != "current/$VM_NAME" ]
+      then
+        continue
+      fi
+    fi
     cd $name
     echo "------------ $name ---- vagrant destroy"
     vagrant destroy -f
